requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.6
colorama==0.4.4
configer @ git+https://github.com/nghorbani/configer@8cd1e3e556d9697298907800a743e120be57ac36
configparser==5.0.2
decorator==4.4.2
freetype-py==2.2.0
future==0.18.2
human-body-prior @ git+https://github.com/totomobile43/human_body_prior.git@821a0e7ebf93f33702babe21c53ecce7d1aa9582
idna==3.2
imageio==2.9.0
imageio-ffmpeg==0.4.5
llvmlite==0.37.0
lmdb==1.2.1
moviepy==1.0.3
networkx==2.6.3
numba==0.54.0
numpy==1.20.3
numpy-quaternion==2021.8.30.10.33.11
opencv-python==4.5.3.56
Pillow==8.3.2
proglog==0.1.9
protobuf==3.18.0
pyglet==1.5.21
PyOpenGL==3.1.0
pyrender==0.1.45
requests==2.26.0
scipy==1.7.1
six==1.16.0
smplx==0.1.28
tabulate==0.8.9
tensorboardX==2.4
torch==1.6.0+cu101
torchgeometry==0.1.2
torchvision==0.7.0+cu101
tqdm==4.62.3
transforms3d==0.3.1
trimesh==3.9.32
urllib3==1.26.7
wincertstore==0.2
